fix bug: 1 mapdatamgr pointer no hold; 2 item effect delay error

+++ mod_room/buff.h
@@ -20,7 +20,7 @@ struct one_effect {
 struct buff_effect {
     uint32_t id;
     struct one_effect effects[BUFF_EFFECT];
-    int time;
+    uint64_t time;
 };
 
 struct delay_vector {
